refactor(dictation): 抽出 dictation_voice_agent 子模块
把 Less Computer 语音 Agent 路径从 coordinator/dictation.rs 机械式搬到
同级 coordinator/dictation_voice_agent.rs（行为保持）：less_computer_approvals、
resolve_less_computer_approval、emit_less_computer、run_voice_agent_transcript、
LessComputerOutcome、run_less_computer_once、maybe_request_approval、
coding_agent_mode_from_pref，以及 LESS_COMPUTER_EVENT/LESS_COMPUTER_APPROVALS。
coordinator.rs glob 重导出新模块并把 resolve_less_computer_approval 调用改到
dictation_voice_agent:: 路径；dictation.rs 去掉随之失用的 AtomicBool 导入。
